Move SwapFactory from actors directory to root of test harness
