+++ api_tests/tests/hbit_herc20.ts
@@ -3,7 +3,7 @@
  * @ledger ethereum
  */
 
-import SwapFactory from "../src/actors/swap_factory";
+import SwapFactory from "../src/swap_factory";
 import { sleep } from "../src/utils";
 import { twoActorTest } from "../src/actor_test";
